move protocols to Protocols.swift file
and generically implement uncompletedCount
