+++ RealmTasks Shared/Protocols.swift
@@ -1,0 +1,38 @@
+/*************************************************************************
+ *
+ * REALM CONFIDENTIAL
+ * __________________
+ *
+ *  [2016] Realm Inc
+ *  All Rights Reserved.
+ *
+ * NOTICE:  All information contained herein is, and remains
+ * the property of Realm Incorporated and its suppliers,
+ * if any.  The intellectual and technical concepts contained
+ * herein are proprietary to Realm Incorporated
+ * and its suppliers and may be covered by U.S. and Foreign Patents,
+ * patents in process, and are protected by trade secret or copyright law.
+ * Dissemination of this information or reproduction of this material
+ * is strictly forbidden unless prior written permission is obtained
+ * from Realm Incorporated.
+ *
+ **************************************************************************/
+
+import RealmSwift
+
+protocol ListPresentable {
+    associatedtype Item: Object, CellPresentable
+    var items: List<Item> { get }
+    var completedCount: Int { get }
+    var uncompletedCount: Int { get }
+}
+
+protocol CellPresentable {
+    var text: String { get set }
+    var completed: Bool { get set }
+    var isCompletable: Bool { get }
+}
+
+extension ListPresentable {
+    var uncompletedCount: Int { return items.count - completedCount }
+}
